+++ src/main/java/com/tp/stage/repository/ClasseRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClasseRepository extends JpaRepository<Classe, Long> {
+public interface ClasseRepository extends JpaRepository<Classe, Integer> {
 
 }
