+++ src/main/java/com/tp/stage/repository/ProfesseurRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfesseurRepository extends JpaRepository<Professeur, Long> {}
+public interface ProfesseurRepository extends JpaRepository<Professeur, Integer> {}
